back button + cancel button + refactor

+++ src/main/java/org/example/model/CurrencyAmount.java
@@ -19,5 +19,5 @@ public class CurrencyAmount {
     @Enumerated(EnumType.STRING)
     private Money currency;
 
-    private Double amount;
+    private long amount;
 }
